Split pytest fixtures into conftest.py
